+++ proto/elys/membershiptier/genesis.proto
@@ -1,14 +1,14 @@
 syntax = "proto3";
 
-package elys.membershiptier;
+package elys.tier;
 
 import "gogoproto/gogo.proto";
-import "elys/membershiptier/params.proto";
-import "elys/membershiptier/portfolio.proto";
+import "elys/tier/params.proto";
+import "elys/tier/portfolio.proto";
 
-option go_package = "github.com/elys-network/elys/x/membershiptier/types";
+option go_package = "github.com/elys-network/elys/x/tier/types";
 
-// GenesisState defines the membershiptier module's genesis state.
+// GenesisState defines the tier module's genesis state.
 message GenesisState {
            Params    params        = 1 [(gogoproto.nullable) = false];
   repeated Portfolio portfolioList = 2 [(gogoproto.nullable) = false];
