+++ proto/elys/membershiptier/portfolio.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
-package elys.membershiptier;
+package elys.tier;
 
-option go_package = "github.com/elys-network/elys/x/membershiptier/types";
+option go_package = "github.com/elys-network/elys/x/tier/types";
 import "gogoproto/gogo.proto"; 
 import "cosmos/base/v1beta1/coin.proto";
