+++ proto/elys/membershiptier/tx.proto
@@ -1,10 +1,10 @@
 syntax = "proto3";
 
-package elys.membershiptier;
+package elys.tier;
 
-import "elys/membershiptier/portfolio.proto";
+import "elys/tier/portfolio.proto";
 
-option go_package = "github.com/elys-network/elys/x/membershiptier/types";
+option go_package = "github.com/elys-network/elys/x/tier/types";
 
 // Msg defines the Msg service.
 service Msg {
