+++ x/tier/genesis.go
@@ -2,8 +2,8 @@ package membershiptier
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/elys-network/elys/x/membershiptier/keeper"
-	"github.com/elys-network/elys/x/membershiptier/types"
+	"github.com/elys-network/elys/x/tier/keeper"
+	"github.com/elys-network/elys/x/tier/types"
 )
 
 // InitGenesis initializes the module's state from a provided genesis state.
