+++ x/tier/keeper/params_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	testkeeper "github.com/elys-network/elys/testutil/keeper"
-	"github.com/elys-network/elys/x/membershiptier/types"
+	"github.com/elys-network/elys/x/tier/types"
 	"github.com/stretchr/testify/require"
 )
