+++ x/tier/keeper/query_portfolio_test.go
@@ -12,7 +12,7 @@ import (
 
 	keepertest "github.com/elys-network/elys/testutil/keeper"
 	"github.com/elys-network/elys/testutil/nullify"
-	"github.com/elys-network/elys/x/membershiptier/types"
+	"github.com/elys-network/elys/x/tier/types"
 )
 
 // Prevent strconv unused error
